backend/requirements.txt
# Python зависимости для SecurityCheck

# Flask и расширения
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0

# База данных
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# HTTP запросы
requests==2.31.0
urllib3==2.1.0

# HTML парсинг
beautifulsoup4==4.12.2
lxml==4.9.3

# SSL/TLS
pyOpenSSL==23.3.0
cryptography==41.0.7

# Stripe платежи
stripe==7.8.0

# AI
openai==1.3.7
anthropic==0.7.7

# Утилиты
python-magic==0.4.27
Pillow==10.1.0
python-whois==0.8.0
dnspython==2.4.2

# Безопасность
Werkzeug==3.0.1
email-validator==2.1.0

# Логирование и мониторинг
python-json-logger==2.0.7

# Тестирование (опционально)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production сервер
gunicorn==21.2.0
